refactor: relocate room purposes module into backend
