+++ src/backend/server/socketGateway.ts
@@ -2,7 +2,7 @@ import type { Server as HttpServer } from 'node:http';
 import { Server as IOServer, type ServerOptions as IOServerOptions, type Socket } from 'socket.io';
 import type { Observable, Subscription } from 'rxjs';
 import { z, type ZodError } from 'zod';
-import type { RoomPurposeSource } from '../../engine/roomPurposes/index.js';
+import type { RoomPurposeSource } from '../src/engine/roomPurposes/index.js';
 import type {
   CommandError,
   CommandResult,
