+++ src/backend/src/engine/environment/deviceDegradation.test.ts
@@ -7,7 +7,7 @@ import type {
   ZoneMetricState,
   ZoneResourceState,
 } from '../../state/models.js';
-import { resolveRoomPurposeId } from '../../../../engine/roomPurposes/index.js';
+import { resolveRoomPurposeId } from '../roomPurposes/index.js';
 import { loadTestRoomPurposes } from '../../testing/loadTestRoomPurposes.js';
 import type { BlueprintRepository } from '../../../data/blueprintRepository.js';
