+++ src/backend/src/engine/index.ts
@@ -10,3 +10,4 @@ export * from './plants/growthModel.js';
 export * from './health/models.js';
 export * from './health/healthEngine.js';
 export * from './workforce/index.js';
+export * from './roomPurposes/index.js';
