+++ src/backend/src/engine/workforce/workforceIntegration.test.ts
@@ -9,7 +9,7 @@ import type {
   ZoneState,
 } from '../../state/models.js';
 import { WorkforceEngine } from './workforceEngine.js';
-import { resolveRoomPurposeId } from '../../../../engine/roomPurposes/index.js';
+import { resolveRoomPurposeId } from '../roomPurposes/index.js';
 import { loadTestRoomPurposes } from '../../testing/loadTestRoomPurposes.js';
 import type { BlueprintRepository } from '../../../data/blueprintRepository.js';
