+++ src/backend/src/sim/loop.test.ts
@@ -1,6 +1,6 @@
 import { beforeAll, describe, expect, it, vi } from 'vitest';
 import { EventBus } from '../lib/eventBus.js';
-import { resolveRoomPurposeId } from '../../../engine/roomPurposes/index.js';
+import { resolveRoomPurposeId } from '../engine/roomPurposes/index.js';
 import { loadTestRoomPurposes } from '../testing/loadTestRoomPurposes.js';
 import type { BlueprintRepository } from '../../data/blueprintRepository.js';
 import type {
