+++ src/backend/src/state/initialization/tasks.test.ts
@@ -3,7 +3,7 @@ import os from 'os';
 import path from 'path';
 import { beforeAll, describe, expect, it } from 'vitest';
 import { RngService } from '../../lib/rng.js';
-import { resolveRoomPurposeId } from '../../../../engine/roomPurposes/index.js';
+import { resolveRoomPurposeId } from '../../engine/roomPurposes/index.js';
 import { loadTestRoomPurposes } from '../../testing/loadTestRoomPurposes.js';
 import type { BlueprintRepository } from '../../../data/blueprintRepository.js';
 import type {
