+++ src/backend/tsconfig.json
@@ -19,7 +19,6 @@
     "data/**/*.ts",
     "facade/**/*.ts",
     "server/**/*.ts",
-    "../engine/**/*.ts",
     "../runtime/**/*.ts"
   ],
   "exclude": ["dist", "node_modules"]
